Modified THcNPSArray
Added tree variables
adcSampPulseNSBInt = Integral (pC) for the NSB samples
adcSampPulseNSAInt = Integral (pC) for the NSA samples
adcSampPulseNumBeforePeak = # of samples between threshold and peak
adcSampPulseNumAfterPeak = # of samples between peak and 10% of peak

Also fixed so that the adcPulseTime and adcSampPulseTime is
filled when the adcCounter > 1080 ( e.g. for scintillators
at adcCounter = 2000, 2001) without the TDCoffset which is only
for the NPS blocks.